renaming and lint
